Conflicts:
	Common/script/cc_error.cpp
	Common/util/textwriter.h
	Engine/ac/character.cpp
	Engine/ac/dynobj/cc_agsdynamicobject.cpp
	Engine/main/game_file.cpp
	Engine/script/ex_global.cpp
	Engine/script/executingscript.h
	Engine/script/script.cpp
	Engine/script/script_runtime.cpp

+++ Plugins/AGSSpriteFont/AGSSpriteFont/SpriteFontRenderer.cpp
@@ -107,15 +107,15 @@ void SpriteFontRenderer::RenderText(const char *text, int fontNumber, BITMAP *de
 void SpriteFontRenderer::Draw(BITMAP *src, BITMAP *dest, int destx, int desty, int srcx, int srcy, int width, int height)
 {
 
-	long srcWidth, srcHeight, destWidth, destHeight, srcColDepth, destColDepth;
+	int srcWidth, srcHeight, destWidth, destHeight, srcColDepth, destColDepth;
 
 	unsigned char **srccharbuffer = _engine->GetRawBitmapSurface (src); //8bit
 	unsigned short **srcshortbuffer = (unsigned short**)srccharbuffer; //16bit;
-    unsigned long **srclongbuffer = (unsigned long**)srccharbuffer; //32bit
+    unsigned int **srclongbuffer = (unsigned int**)srccharbuffer; //32bit
 
 	unsigned char **destcharbuffer = _engine->GetRawBitmapSurface (dest); //8bit
 	unsigned short **destshortbuffer = (unsigned short**)destcharbuffer; //16bit;
-    unsigned long **destlongbuffer = (unsigned long**)destcharbuffer; //32bit
+    unsigned int **destlongbuffer = (unsigned int**)destcharbuffer; //32bit
 
 	int transColor = _engine->GetBitmapTransparentColor(src);
